Entities refactored source root.
