--- src/main/entities/UserFactoryInterface.java
@@ -1,5 +1,0 @@
-package entities;
-
-public interface UserFactoryInterface {
-    User create(String type, String username, String name);
-}
