+++ src/main/java/entities/UserFactoryInterface.java
@@ -1,0 +1,5 @@
+package entities;
+
+public interface UserFactoryInterface {
+    User create(String type, String username, String name);
+}
